back, front: move contact email code in the backend
We set up a service that allows users to contact us via email through our contact page. Initially, the external email service was called directly from the frontend.
However, this approach was insecure, as any user could potentially extract the credentials and misuse them.
To address this, the call to the external email service is now handled securely from the backend.

Signed-off-by: Clara Ni <clara.ni@outlook.fr>